rewire toplayer
